requirements.txt
openai==1.78.1
python-dotenv==1.1.0
PyYAML==6.0.1
